Clean up features and UI (#253)
* Replace service title with the document title

* Add subpages to pages instead of groups

* Add milkdown page(experimental)

---------
