+++ package.json
@@ -14,9 +14,23 @@
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
     "@marp-team/marpit": "^2.4.2",
+    "@milkdown/core": "^7.1.1",
+    "@milkdown/ctx": "^7.1.1",
+    "@milkdown/plugin-block": "^7.1.1",
+    "@milkdown/plugin-cursor": "^7.1.1",
+    "@milkdown/plugin-history": "^7.1.1",
+    "@milkdown/plugin-listener": "^7.1.1",
+    "@milkdown/plugin-slash": "^7.1.1",
+    "@milkdown/preset-commonmark": "^7.1.1",
+    "@milkdown/prose": "^7.1.1",
+    "@milkdown/react": "^7.1.1",
+    "@milkdown/theme-nord": "^7.1.1",
+    "@milkdown/transformer": "^7.1.1",
+    "@milkdown/utils": "^7.1.1",
     "@mui/icons-material": "^5.11.11",
     "@mui/lab": "^5.0.0-alpha.123",
     "@mui/material": "^5.11.13",
+    "@prosemirror-adapter/react": "^0.2.4",
     "@react-hook/throttle": "^2.2.0",
     "@reduxjs/toolkit": "^1.9.3",
     "@sentry/react": "^7.43.0",
