+++ src/components/Editor/mime/text/milkdown/Editor.scss
@@ -1,0 +1,22 @@
+.milkdown-theme-editor {
+  background-color: yellow;
+  height: calc(100vh - 80px);
+  outline: none;
+}
+
+.milkdown {
+  background-color: slategray;
+  border: 1px solid slategray;
+  border-radius: 5px;
+  padding: 4px 2px;
+  margin: 5px;
+
+  .ProseMirror {
+    outline: none;
+  }
+}
+
+.editor {
+  width: 100%;
+  height: 100%;
+}
