+++ src/constants/editor.ts
@@ -1,1 +1,10 @@
 export const NAVBAR_HEIGHT = 90;
+
+export enum MimeType {
+  MARKDOWN = 'text/markdown',
+  MILKDOWN = 'text/milkdown',
+  PLAIN = 'text/plain',
+  WHITEBOARD = 'application/vnd.pairy.whiteboard',
+  CELL = 'application/cell',
+  JSON = 'application/json',
+}
